chore(labels): move and rename label files
