+++ src/app/Recordings/ArchivedRecordingsTable.tsx
@@ -49,7 +49,7 @@ import { Observable, forkJoin, merge, combineLatest } from 'rxjs';
 import { concatMap, filter, first, map } from 'rxjs/operators';
 import { NO_TARGET } from '@app/Shared/Services/Target.service';
 import { parseLabels } from '@app/RecordingMetadata/RecordingLabel';
-import { EditableLabelCell } from './EditableLabelCell';
+import { EditableLabelCell } from '../RecordingMetadata/EditableLabelCell';
 import { BulkEditLabels } from '@app/RecordingMetadata/BulkEditLabels';
 
 export interface ArchivedRecordingsTableProps { }
